examples/requirements.txt
# OpenAI Agents SDK and dependencies
agents>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0

# Rich console output
rich>=13.0.0
click>=8.0.0

# HTTP requests for testing
requests>=2.28.0

# Path handling
pathlib2>=2.3.0; python_version < "3.4" 